Conflicts:
	android/cli/commands/_build.js
	android/modules/ui/src/java/ti/modules/titanium/ui/TabGroupProxy.java
	android/modules/ui/src/java/ti/modules/titanium/ui/android/AndroidModule.java
	android/modules/ui/src/java/ti/modules/titanium/ui/widget/TiUILabel.java
	android/modules/ui/src/java/ti/modules/titanium/ui/widget/TiUISwitch.java
	android/modules/ui/src/java/ti/modules/titanium/ui/widget/listview/ViewItem.java
	android/modules/ui/src/java/ti/modules/titanium/ui/widget/tabgroup/TiUIActionBarTabGroup.java
	android/titanium/src/java/org/appcelerator/titanium/TiBaseActivity.java
	android/titanium/src/java/org/appcelerator/titanium/util/TiUIHelper.java
	iphone/Classes/APIModule.m
	iphone/Classes/KrollObject.m
	iphone/Classes/MediaModule.m
	iphone/Classes/TiAnimation.m
	iphone/Classes/TiMediaAudioPlayerProxy.h
	iphone/Classes/TiMediaAudioRecorderProxy.h
	iphone/Classes/TiMediaAudioSession.h
	iphone/Classes/TiMediaMusicPlayer.h
	iphone/Classes/TiMediaSoundProxy.h
	iphone/Classes/TiNetworkHTTPClientProxy.m
	iphone/Classes/TiRootViewController.m
	iphone/Classes/TiUILabel.h
	iphone/Classes/TiUIListView.m
	iphone/Classes/TiUIScrollableView.h
	iphone/Classes/TiUIScrollableView.m
	iphone/Classes/TiUIiOSNavWindow.h
	iphone/Classes/TiUIiPhoneActivityIndicatorStyleProxy.h
	iphone/Classes/TiUIiPhoneTableViewSeparatorStyleProxy.h
	iphone/Classes/TiUtils.m
	iphone/Classes/TiViewProxy.m
	iphone/cli/commands/_build.js
	iphone/iphone/Titanium.xcodeproj/project.pbxproj
	support/module/android/build.xml

+++ iphone/Classes/TiNetworkSocketTCPProxy.m
@@ -829,7 +829,7 @@ TYPESAFE_SETTER(setError, error, KrollCallback)
             case TO_STREAM: {
                 // Perform the write to stream
                 id<TiStreamInternal> stream = [info valueForKey:@"destination"];
-                int size = [TiUtils intValue:[info valueForKey:@"chunkSize"]];
+                NSInteger size = [TiUtils intValue:[info valueForKey:@"chunkSize"]];
                 KrollCallback* callback = [info valueForKey:@"callback"];
                 
                 TiBuffer* tempBuffer = [[[TiBuffer alloc] _initWithPageContext:[self executionContext]] autorelease];
@@ -846,7 +846,7 @@ TYPESAFE_SETTER(setError, error, KrollCallback)
             case TO_CALLBACK: {
                 // Perform the pump to callback
                 KrollCallback* callback = [info valueForKey:@"callback"];
-                int size = [TiUtils intValue:[info valueForKey:@"chunkSize"]];
+                NSInteger size = [TiUtils intValue:[info valueForKey:@"chunkSize"]];
                 
                 TiBuffer* tempBuffer = [[[TiBuffer alloc] _initWithPageContext:[self executionContext]] autorelease];
                 [tempBuffer setData:[NSMutableData dataWithData:data]];
